requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
python-multipart